requirements.txt
alabaster==0.7.8
ansicolors==1.0.2
Babel==2.3.4
backports.shutil-get-terminal-size==1.0.0
configparser==3.5.0b2
coverage==4.0.1
decorator==4.0.9
dj-database-url==0.4.1
Django==1.9.2
django-cors-headers==1.1.0
django-debug-toolbar==1.4
django-environ==0.4.0
django-extensions==1.6.1
django-filter==0.11.0
django-gulp==2.0.0
django-model-utils==2.3.1
django-secure==1.0.1
django-test-plus==1.0.9
django-thumbs==0.4
djangorestframework==3.3.2
docutils==0.12
drf-extensions==0.2.8
easy-thumbnails==2.3
env-tools==2.0.0
factory-boy==2.5.2
futures==2.2.0
ipdb==0.8.1
ipython==4.2.0
ipython-genutils==0.1.0
Jinja2==2.8
MarkupSafe==0.23
mpmath==0.19
pathlib2==2.1.0
pexpect==4.0.1
pickleshare==0.7.2
Pillow==3.1.1
psutil==2.2.1
psycopg2==2.6.1
ptyprocess==0.5.1
Pygments==2.1.3
pytz==2015.6
requests==2.9.1
simplegeneric==0.8.1
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.3.1
sphinx-rtd-theme==0.1.9
sqlparse==0.1.19
sympy==1.0
tini==3.0.1
traitlets==4.2.1
Werkzeug==0.10.4
whitenoise==2.0.4
